Hack to fix runtime assertion in MSVC debug mode
Hmmm, my reading of the standard (okay, cppreference) suggests that
comparison again a value-initialised iterator should be okay, but
apparently MSVCs debug iterators don't like it, so let's shut them up.

Sadly this requires that Sentinel is convertible to InputIt, which
rather goes against the point of using a separate sentinel type in the
first place. Better fix required.